requirements.txt
torch
torchvision
streamlit
Pillow
tqdm
matplotlib